add: Cap02 exercises

+++ 02-Variables_and_Primitive_Data_Types/exercises/Exercise2_1.java
@@ -3,12 +3,12 @@ package exercises;
 public class Exercise2_1 {
     public static void main(String[] args) {
         String breed;
-        float height, weight, length;
+        double height, weight, length;
 
         breed = "Velociraptor";
-        height = 0.5f;
-        weight = 16.5f;
-        length = 1.75f;
+        height = 0.5;
+        weight = 16.5;
+        length = 1.75;
 
         System.out.println("Breed: " + breed + "\nHeight: " + height + "m" + "\nWeight: " + weight + "kg" + "\nLength: "
                 + length + "m");
